ChatBot/requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
supabase==2.3.3
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
faiss-cpu==1.7.4
unstructured==0.10.30
sentence-transformers==2.2.2
pypdf==3.17.4
python-docx==1.1.0

langchain==0.0.335
langchain-core==0.0.10
langchain-community==0.0.2
pydantic==1.10.12

langchain-google-genai==0.0.5
schedule==1.2.2

gtts==2.5.4
pyttsx3==2.99
SpeechRecognition==3.14.3